REFACTOR country service to be within it own folder
